Moved all the new components to learner dashboard folder
