+++ core/templates/pages/learner-dashboard-page/community-lessons-tab.component.spec.ts
