+++ core/templates/pages/learner-dashboard-page/goals-tab.component.html
