+++ core/templates/pages/learner-dashboard-page/learner-dashboard-page.module.ts
@@ -25,11 +25,11 @@ import { RequestInterceptor } from 'services/request-interceptor.service';
 import { SharedComponentsModule } from 'components/shared-component.module';
 
 import { LearnerStorySummaryTileComponent } from 'components/summary-tile/learner-story-summary-tile.component';
-import { ProgressTabComponent } from './new-learner-dashboard/progress-tab.component';
-import { GoalsTabComponent } from './new-learner-dashboard-page/goals-tab.component';
-import { CommunityLessonsTabComponent } from './new-learner-dashboard/community-lessons-tab.component';
+import { ProgressTabComponent } from './progress-tab.component';
+import { GoalsTabComponent } from './goals-tab.component';
+import { CommunityLessonsTabComponent } from './community-lessons-tab.component';
 import { LearnerTopicSummaryTileComponent } from 'components/summary-tile/learner-topic-summary-tile.component';
-import { HomeTabComponent } from './new-learner-dashboard-page/home-tab.component';
+import { HomeTabComponent } from './home-tab.component';
 import { LearnerDashboardPageComponent } from './learner-dashboard-page.component';
 import { OppiaAngularRootComponent } from 'components/oppia-angular-root.component';
 import { platformFeatureInitFactory, PlatformFeatureService } from 'services/platform-feature.service';
